+++ query/generator.go
@@ -1,0 +1,63 @@
+package query
+
+import (
+	"context"
+
+	"github.com/dgraph-io/dgraph/algo"
+	"github.com/dgraph-io/dgraph/tok"
+	"github.com/dgraph-io/dgraph/x"
+)
+
+func allof(ctx context.Context, intersectUIDs *algo.UIDList, attr, terms string) (*algo.UIDList, error) {
+	sg := &SubGraph{Attr: attr}
+	sgChan := make(chan error, 1)
+
+	// Tokenize FuncArgs[1].
+	tokenizer, err := tok.NewTokenizer([]byte(terms))
+	if err != nil {
+		return nil, x.Errorf("Could not create tokenizer: %v", terms)
+	}
+	defer tokenizer.Destroy()
+	x.AssertTrue(tokenizer != nil)
+	tokens := tokenizer.Tokens()
+	taskQuery := createTaskQuery(sg, nil, tokens, intersectUIDs)
+	go ProcessGraph(ctx, sg, taskQuery, sgChan)
+	select {
+	case <-ctx.Done():
+		return nil, x.Wrap(ctx.Err())
+	case err = <-sgChan:
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	x.AssertTrue(len(sg.Result) == len(tokens))
+	return algo.IntersectLists(sg.Result), nil
+}
+
+func anyof(ctx context.Context, intersectUIDs *algo.UIDList, attr, terms string) (*algo.UIDList, error) {
+	sg := &SubGraph{Attr: attr}
+	sgChan := make(chan error, 1)
+
+	// Tokenize FuncArgs[1].
+	tokenizer, err := tok.NewTokenizer([]byte(terms))
+	if err != nil {
+		return nil, x.Errorf("Could not create tokenizer: %v", terms)
+	}
+	defer tokenizer.Destroy()
+	x.AssertTrue(tokenizer != nil)
+	tokens := tokenizer.Tokens()
+	taskQuery := createTaskQuery(sg, nil, tokens, intersectUIDs)
+	go ProcessGraph(ctx, sg, taskQuery, sgChan)
+	select {
+	case <-ctx.Done():
+		return nil, x.Wrap(ctx.Err())
+	case err = <-sgChan:
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	x.AssertTrue(len(sg.Result) == len(tokens))
+	return algo.MergeLists(sg.Result), nil
+}
